Bachelors final year project
This project aims to educate the user on the measures one should take in the practical
world to stay safe from various cyberthreats. Hence the game is divided into 2 levels.
For the first level, the user plays a solitaire-like card game against the system so that
he/she gets accustomed to the different cyberthreats and what defense measures are
taken against them. The user will be tracked on the number of points. After reaching a
benchmark number of points, the user gets access to the Level 2. Here, considering that
the user is familiar with the concept of the game, plays an Attack-Defense game with
another player in real time where they get assigned roles of an Attacker or a Defender
on turn-by-turn basis. So essentially, this level is a test of knowledge for the user.